+++ CHANGES.rst
@@ -5,6 +5,8 @@ Version 3.0.5
 
 Unreleased
 
+-   The Watchdog reloader ignores file closed no write events. :issue:`2945`
+
 
 Version 3.0.4
 -------------